Rename internal formaters to include vendor name
