+++ Tests/Blazorise.Tests/Helpers/BlazoriseConfig.cs
@@ -6,6 +6,8 @@ using Blazorise.Licensing;
 using Blazorise.Localization;
 using Blazorise.Modules;
 using Blazorise.Utilities;
+using Blazorise.Utilities.Vendors;
+using Blazorise.Vendors;
 using Bunit;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.JSInterop;
@@ -30,8 +32,8 @@ public static class BlazoriseConfig
         services.AddSingleton<ValidatorValidationHandler>();
         services.AddSingleton<PatternValidationHandler>();
         services.AddSingleton<DataAnnotationValidationHandler>();
-        services.AddSingleton<IDateTimeDisplayFormatConverter, DateTimeDisplayFormatConverter>();
-        services.AddSingleton<IDateTimeInputFormatConverter, DateTimeInputFormatConverter>();
+        services.AddSingleton<IFlatPickrDateTimeDisplayFormatConverter, FlatPickrDateTimeDisplayFormatConverter>();
+        services.AddSingleton<IInputMaskDateTimeInputFormatConverter, InputMaskDateTimeInputFormatConverter>();
         services.AddSingleton<IVersionProvider, VersionProvider>();
         services.AddScoped<ITextLocalizerService, TextLocalizerService>();
         services.AddScoped( typeof( ITextLocalizer<> ), typeof( TextLocalizer<> ) );
